1. Add checkApplication() for NestLedger, and add ApplicationChanged event 2. Delete the logic that the eaten list can be retrieved in advance to avoid being used 3. Change INestMining.Config.maxBiteNestedLevel to uint8

+++ contracts/interface/INestMining.sol
@@ -39,7 +39,7 @@ interface INestMining {
 
         // -- Public configuration
         // The number of times the sheet assets have doubled. 4
-        uint16 maxBiteNestedLevel;
+        uint8 maxBiteNestedLevel;
         
         // Price effective block interval. 20
         uint16 priceEffectSpan;
